# infrastructure/openapi.yaml
swagger: '2.0'
info:
  title: F1 Visualizer API
  description: The central API Gateway routing traffic to F1V microservices.
  version: 1.0.0
schemes:
  - https
produces:
  - application/json
consumes:
  - application/json
securityDefinitions:
  auth0_jwt:
    authorizationUrl: "AUTH0_ISSUER_URI_PLACEHOLDER/authorize"
    flow: "implicit"
    type: "oauth2"
    x-google-issuer: "AUTH0_ISSUER_URI_PLACEHOLDER/"
    x-google-jwks_uri: "AUTH0_ISSUER_URI_PLACEHOLDER/.well-known/jwks.json"
    x-google-audiences: "AUTH0_AUDIENCE_PLACEHOLDER"
paths:
  # --- USER SERVICE ---
  /api/v1/users/me:
    get:
      summary: Get Current User
      operationId: getUser
      x-google-backend:
        address: USER_SERVICE_URL_PLACEHOLDER
        path_translation: APPEND_PATH_TO_ADDRESS
        deadline: 60.0
      responses:
        '200':
          description: OK
      security:
        - auth0_jwt: []
    options:
      summary: CORS support
      operationId: userOptionsMe
      x-google-backend:
        address: USER_SERVICE_URL_PLACEHOLDER
        path_translation: APPEND_PATH_TO_ADDRESS
        deadline: 60.0
      responses:
        '200':
          description: OK
          headers:
            Access-Control-Allow-Origin:
              type: string
            Access-Control-Allow-Methods:
              type: string
            Access-Control-Allow-Headers:
              type: string
            Access-Control-Allow-Credentials:
              type: string
  /api/v1/users/me/preferences:
    put:
      summary: Update Preferences
      operationId: updateUserPrefs
      x-google-backend:
        address: USER_SERVICE_URL_PLACEHOLDER
        path_translation: APPEND_PATH_TO_ADDRESS
        deadline: 60.0
      responses:
        '200':
          description: OK
      security:
        - auth0_jwt: []
    options:
      summary: CORS support
      operationId: userOptionsPrefs
      x-google-backend:
        address: USER_SERVICE_URL_PLACEHOLDER
        path_translation: APPEND_PATH_TO_ADDRESS
        deadline: 60.0
      responses:
        '200':
          description: OK
          headers:
            Access-Control-Allow-Origin:
              type: string
            Access-Control-Allow-Methods:
              type: string
            Access-Control-Allow-Headers:
              type: string
            Access-Control-Allow-Credentials:
              type: string

  # --- ANALYSIS SERVICE ---
  /api/v1/analysis/session/{sessionKey}/laps:
    get:
      summary: Get Laps
      operationId: getLaps
      parameters:
        - name: sessionKey
          in: path
          required: true
          type: string
      x-google-backend:
        address: DATA_ANALYSIS_SERVICE_URL_PLACEHOLDER
        path_translation: APPEND_PATH_TO_ADDRESS
        deadline: 60.0
      responses:
        '200':
          description: OK
      security:
        - auth0_jwt: []
    options:
      summary: CORS support
      operationId: analysisOptions
      parameters:
        - name: sessionKey
          in: path
          required: true
          type: string
      x-google-backend:
        address: DATA_ANALYSIS_SERVICE_URL_PLACEHOLDER
        path_translation: APPEND_PATH_TO_ADDRESS
        deadline: 60.0
      responses:
        '200':
          description: OK
          headers:
            Access-Control-Allow-Origin:
              type: string
            Access-Control-Allow-Methods:
              type: string
            Access-Control-Allow-Headers:
              type: string
            Access-Control-Allow-Credentials:
              type: string
  /api/v1/analysis/sessions:
    get:
      summary: Get Sessions
      operationId: getSessions
      x-google-backend:
        address: DATA_ANALYSIS_SERVICE_URL_PLACEHOLDER
        path_translation: APPEND_PATH_TO_ADDRESS
        deadline: 60.0
      responses:
        '200':
          description: OK
      security:
        - auth0_jwt: [ ]
    options:
      summary: CORS support
      operationId: sessionsOptions
      x-google-backend:
        address: DATA_ANALYSIS_SERVICE_URL_PLACEHOLDER
        path_translation: APPEND_PATH_TO_ADDRESS
        deadline: 60.0
      responses:
        '200':
          description: OK
          headers:
            Access-Control-Allow-Origin:
              type: string
            Access-Control-Allow-Methods:
              type: string
            Access-Control-Allow-Headers:
              type: string
            Access-Control-Allow-Credentials:
              type: string
  /api/v1/analysis/drivers:
    get:
      summary: Get Drivers
      operationId: getDrivers
      x-google-backend:
        address: DATA_ANALYSIS_SERVICE_URL_PLACEHOLDER
        path_translation: APPEND_PATH_TO_ADDRESS
        deadline: 60.0
      responses:
        '200':
          description: OK
      security:
        - auth0_jwt: [ ]
    options:
      summary: CORS support
      operationId: driversOptions
      x-google-backend:
        address: DATA_ANALYSIS_SERVICE_URL_PLACEHOLDER
        path_translation: APPEND_PATH_TO_ADDRESS
        deadline: 60.0
      responses:
        '200':
          description: OK
          headers:
            Access-Control-Allow-Origin:
              type: string
            Access-Control-Allow-Methods:
              type: string
            Access-Control-Allow-Headers:
              type: string
            Access-Control-Allow-Credentials:
              type: string
  /api/v1/analysis/sessions/search:
    get:
      summary: Search Sessions
      operationId: searchSessions
      parameters:
        - name: query
          in: query
          required: false
          type: string
      x-google-backend:
        address: DATA_ANALYSIS_SERVICE_URL_PLACEHOLDER
        path_translation: APPEND_PATH_TO_ADDRESS
        deadline: 60.0
      responses:
        '200':
          description: OK
      security:
        - auth0_jwt: [ ]
    options:
      summary: CORS support
      operationId: searchSessionsOptions
      x-google-backend:
        address: DATA_ANALYSIS_SERVICE_URL_PLACEHOLDER
        path_translation: APPEND_PATH_TO_ADDRESS
        deadline: 60.0
      responses:
        '200':
          description: OK
          headers:
            Access-Control-Allow-Origin:
              type: string
            Access-Control-Allow-Methods:
              type: string
            Access-Control-Allow-Headers:
              type: string
            Access-Control-Allow-Credentials:
              type: string
  /api/v1/analysis/drivers/{driverId}/stats:
    get:
      summary: Get Driver Stats
      operationId: getDriverStats
      parameters:
        - name: driverId
          in: path
          required: true
          type: integer
      x-google-backend:
        address: DATA_ANALYSIS_SERVICE_URL_PLACEHOLDER
        path_translation: APPEND_PATH_TO_ADDRESS
        deadline: 60.0
      responses:
        '200':
          description: OK
      security:
        - auth0_jwt: [ ]
    options:
      summary: CORS support
      operationId: driverStatsOptions
      parameters:
        - name: driverId
          in: path
          required: true
          type: integer
      x-google-backend:
        address: DATA_ANALYSIS_SERVICE_URL_PLACEHOLDER
        path_translation: APPEND_PATH_TO_ADDRESS
        deadline: 60.0
      responses:
        '200':
          description: OK
          headers:
            Access-Control-Allow-Origin:
              type: string
            Access-Control-Allow-Methods:
              type: string
            Access-Control-Allow-Headers:
              type: string
            Access-Control-Allow-Credentials:
              type: string
  # --- INGESTION SERVICE ---
  /api/v1/ingestion/command:
    post:
      summary: Issue Command
      operationId: ingestionCommand
      x-google-backend:
        address: DATA_INGESTION_SERVICE_URL_PLACEHOLDER
        path_translation: APPEND_PATH_TO_ADDRESS
        deadline: 60.0
      responses:
        '200':
          description: OK
      security:
        - auth0_jwt: []
    options:
      summary: CORS support
      operationId: ingestionOptionsCommand
      x-google-backend:
        address: DATA_INGESTION_SERVICE_URL_PLACEHOLDER
        path_translation: APPEND_PATH_TO_ADDRESS
        deadline: 60.0
      responses:
        '200':
          description: OK
          headers:
            Access-Control-Allow-Origin:
              type: string
            Access-Control-Allow-Methods:
              type: string
            Access-Control-Allow-Headers:
              type: string
            Access-Control-Allow-Credentials:
              type: string
  /api/v1/ingestion/load-historical:
    post:
      summary: Load Historical Data
      operationId: loadHistorical
      x-google-backend:
        address: DATA_INGESTION_SERVICE_URL_PLACEHOLDER
        path_translation: APPEND_PATH_TO_ADDRESS
        deadline: 120.0
      responses:
        '200':
          description: OK
      security:
        - auth0_jwt: []
    options:
      summary: CORS support
      operationId: ingestionOptionsLoad
      x-google-backend:
        address: DATA_INGESTION_SERVICE_URL_PLACEHOLDER
        path_translation: APPEND_PATH_TO_ADDRESS
        deadline: 60.0
      responses:
        '200':
          description: OK
          headers:
            Access-Control-Allow-Origin:
              type: string
            Access-Control-Allow-Methods:
              type: string
            Access-Control-Allow-Headers:
              type: string
            Access-Control-Allow-Credentials:
              type: string
  /api/v1/ingestion/load-reference:
    post:
      summary: Load Reference Data
      operationId: loadReference
      parameters:
        - name: year
          in: query
          required: false
          type: integer
      x-google-backend:
        address: DATA_INGESTION_SERVICE_URL_PLACEHOLDER
        path_translation: APPEND_PATH_TO_ADDRESS
        deadline: 120.0
      responses:
        '200':
          description: OK
      security:
        - auth0_jwt: []
    options:
      summary: CORS support
      operationId: ingestionOptionsReference
      x-google-backend:
        address: DATA_INGESTION_SERVICE_URL_PLACEHOLDER
        path_translation: APPEND_PATH_TO_ADDRESS
        deadline: 60.0
      responses:
        '200':
          description: OK
          headers:
            Access-Control-Allow-Origin:
              type: string
            Access-Control-Allow-Methods:
              type: string
            Access-Control-Allow-Headers:
              type: string
            Access-Control-Allow-Credentials:
              type: string
  /api/v1/ingestion/playback/play:
    post:
      summary: Play Simulation
      operationId: playSimulation
      x-google-backend:
        address: DATA_INGESTION_SERVICE_URL_PLACEHOLDER
        path_translation: APPEND_PATH_TO_ADDRESS
        deadline: 60.0
      responses:
        '200':
          description: OK
      security:
        - auth0_jwt: [ ]
    options:
      summary: CORS support
      operationId: playOptions
      x-google-backend:
        address: DATA_INGESTION_SERVICE_URL_PLACEHOLDER
        path_translation: APPEND_PATH_TO_ADDRESS
        deadline: 60.0
      responses:
        '200':
          description: OK
          headers:
            Access-Control-Allow-Origin:
              type: string
            Access-Control-Allow-Methods:
              type: string
            Access-Control-Allow-Headers:
              type: string
            Access-Control-Allow-Credentials:
              type: string
  /api/v1/ingestion/playback/pause:
    post:
      summary: Pause Simulation
      operationId: pauseSimulation
      x-google-backend:
        address: DATA_INGESTION_SERVICE_URL_PLACEHOLDER
        path_translation: APPEND_PATH_TO_ADDRESS
        deadline: 60.0
      responses:
        '200':
          description: OK
      security:
        - auth0_jwt: [ ]
    options:
      summary: CORS support
      operationId: pauseOptions
      x-google-backend:
        address: DATA_INGESTION_SERVICE_URL_PLACEHOLDER
        path_translation: APPEND_PATH_TO_ADDRESS
        deadline: 60.0
      responses:
        '200':
          description: OK
          headers:
            Access-Control-Allow-Origin:
              type: string
            Access-Control-Allow-Methods:
              type: string
            Access-Control-Allow-Headers:
              type: string
            Access-Control-Allow-Credentials:
              type: string
  /api/v1/ingestion/playback/seek:
    post:
      summary: Seek Simulation
      operationId: seekSimulation
      parameters:
        - name: percentage
          in: query
          required: true
          type: integer
      x-google-backend:
        address: DATA_INGESTION_SERVICE_URL_PLACEHOLDER
        path_translation: APPEND_PATH_TO_ADDRESS
        deadline: 60.0
      responses:
        '200':
          description: OK
      security:
        - auth0_jwt: [ ]
    options:
      summary: CORS support
      operationId: seekOptions
      x-google-backend:
        address: DATA_INGESTION_SERVICE_URL_PLACEHOLDER
        path_translation: APPEND_PATH_TO_ADDRESS
        deadline: 60.0
      responses:
        '200':
          description: OK
          headers:
            Access-Control-Allow-Origin:
              type: string
            Access-Control-Allow-Methods:
              type: string
            Access-Control-Allow-Headers:
              type: string
            Access-Control-Allow-Credentials:
              type: string
  # --- TELEMETRY SERVICE ---
  /api/v1/debug/publish:
    post:
      summary: Publish Debug Telemetry
      operationId: debugPublish
      x-google-backend:
        address: TELEMETRY_SERVICE_URL_PLACEHOLDER
        path_translation: APPEND_PATH_TO_ADDRESS
        deadline: 60.0
      responses:
        '200':
          description: OK
      security:
        - auth0_jwt: [ ]
    options:
      summary: CORS support
      operationId: debugPublishOptions
      x-google-backend:
        address: TELEMETRY_SERVICE_URL_PLACEHOLDER
        path_translation: APPEND_PATH_TO_ADDRESS
        deadline: 60.0
      responses:
        '200':
          description: OK
          headers:
            Access-Control-Allow-Origin:
              type: string
            Access-Control-Allow-Methods:
              type: string
            Access-Control-Allow-Headers:
              type: string
            Access-Control-Allow-Credentials:
              type: string